Solvers: move our solvers to the API.
