+++ src/api/libopencor/module/libopencor
@@ -20,6 +20,14 @@ limitations under the License.
 #include "libopencor/issue.h"
 #include "libopencor/logger.h"
 #include "libopencor/solver.h"
+#include "libopencor/solvercvode.h"
+#include "libopencor/solverforwardeuler.h"
+#include "libopencor/solverfourthorderrungekutta.h"
+#include "libopencor/solverheun.h"
 #include "libopencor/solverinfo.h"
+#include "libopencor/solverkinsol.h"
+#include "libopencor/solvernla.h"
+#include "libopencor/solverode.h"
 #include "libopencor/solverproperty.h"
+#include "libopencor/solversecondorderrungekutta.h"
 #include "libopencor/version.h"
