+++ src/solver/solverforwardeuler_p.h
@@ -16,8 +16,9 @@ limitations under the License.
 
 #pragma once
 
-#include "solverforwardeuler.h"
 #include "solverode_p.h"
+
+#include "libopencor/solverforwardeuler.h"
 
 namespace libOpenCOR {
