+++ tests/api/solver/cvodetests.cpp
@@ -15,7 +15,6 @@ limitations under the License.
 */
 
 #include "odemodel.h"
-#include "solvercvode.h"
 #include "solvers.h"
 #include "utils.h"
