+++ tests/api/solver/fourthorderrungekuttatests.cpp
@@ -15,7 +15,6 @@ limitations under the License.
 */
 
 #include "odemodel.h"
-#include "solverfourthorderrungekutta.h"
 #include "solvers.h"
 #include "utils.h"
