+++ tests/api/solver/nlamodel.h
@@ -16,7 +16,7 @@ limitations under the License.
 
 #pragma once
 
-#include "solvernla.h"
+#include "libopencor/solvernla.h"
 
 #include "tests/utils.h"
